Add new phases in the font generation - first we generate the font from the existing glyphs (pt4), then we add derived characters (pt5), tweak the properties (pt6), then generate all of the formats (pt7)
