+++ xkcd-script/generator/run.sh
@@ -12,4 +12,6 @@ $RUN_CTXT python3 pt1_character_extraction.py
 $RUN_CTXT python3 pt2_character_classification.py
 $RUN_CTXT python3 pt3_ppm_to_svg.py
 $RUN_CTXT python3 pt4_svg_to_font.py
-$RUN_CTXT python3 pt5_gen_reprod_font.py
+$RUN_CTXT python3 pt5_derived_chars.py
+$RUN_CTXT python3 pt6_font_properties.py
+$RUN_CTXT python3 pt7_gen_reprod_font.py
